remove lives in gcp and prometheus
