+++ .nycrc
@@ -10,8 +10,6 @@
     "lib/metrics-aws.js",
     "lib/metrics-gcp.js",
     "lib/metrics.js",
-    "lib/metricsHelper.js",
-    "lib/prometheus.js",
     "test/*"
   ]
 }
