+++ lib/routes.js
@@ -2,11 +2,9 @@
 
 const express = require("express");
 const alive = require("./routes/alive.js");
-const prometheus = require("./routes/prometheus");
 
 const router = express.Router(); // eslint-disable-line new-cap
 router.get("/_alive", alive);
 router.head("/_alive", alive);
-router.get("/metrics/", prometheus.metrics);
 
 module.exports = router;
