+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lu-server",
-  "version": "5.1.1",
+  "version": "6.0.0",
   "engines": {
     "node": ">=16 <=18",
     "yarn": "please use npm"
@@ -31,7 +31,6 @@
     "express": "^4.18.2",
     "joi": "^17.7.0",
     "lu-logger": "github:BonnierNews/lu-logger#semver:^7.1.0",
-    "prom-client": "^12.0.0",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
